requirements.txt
Django>=2.1.5,<2.2.0
djangorestframework>=3.9.1,<4.0.0
flake8>=3.6.0,<3.7.0
psycopg2>=2.7.5,<3
